+++ lib/src/voyager_widget.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Router;
 import 'package:provider/provider.dart';
 import 'package:voyager/voyager.dart';